fix: Improve error handling when monitored appliance is not available
Previously, the program yielded the rather cryptic error message
`<built-in function get> returned NULL without setting an error` and
also added a traceback, not catching the exception.

Now, any such exceptions are caught and cleanly converged into the
`UNKNOWN` state, so that the program croaks with:

    UNKNOWN - timed out while connecting to remote host

The improved error message will only be available when installing the
latest `easysnmp` from the upstream repository.